[ql-tipc] Fail only on negative mmc_rpmb_response
mmc_rpmb_response will return negative if the RPMB read did not execute
A non-negative return value is the "result" field in the read RPMB data
frame. Trusty's storage service expects to receive an RPMB data frame,
even if the "result" field indicates an error, so we should not return
an error when mmc_rpmb_response returns positive.

Change-Id: I83fd67567dc9deb0bb328bcf791e9f31fa586744