ql-tipc: SMP support
Add trusty_dev_nop that secondary CPUs can call to enter trusty. Add an
argument to the idle callbaks to specify that it is called after
trusty_ipc_poll_for_event returned TRUSTY_EVENT_NONE. In this case the
the idle call needs to abort if a secondary CPU (any CPU other than the
CPU that is waiting for a message) has entered idle since the last such
call.

Bug: 116851819
Change-Id: I59579209e9eb52be345b1284f89676810ab425f0